Reorganized parsers.py
